### 2017-08-24 #376 @cMadan
@raamana - I'd be happy to serve as editor for this submission

@KirstieJane, @oesteban, @JohnGriffiths - Are any of you interested/available to be a reviewer for this submission? For an overview of the review process, please see the reviewer guidelines (http://joss.theoj.org/about#reviewer_guidelines)
