### 2017-08-24 #376 @oesteban
@cMadan @raamana I'd very much like to review this submission :)
