### 2017-08-24 29533f7
Citations to scipy, numpy and medpy [skip ci]
