+++ paper.bib
@@ -9,6 +9,34 @@
 	eprint = {http://www.biorxiv.org/content/early/2017/07/31/170381.full.pdf},
 	journal = {bioRxiv}
 }
+@article{oliphant2007python,
+  title={Python for scientific computing},
+  author={Oliphant, Travis E},
+  journal={Computing in Science \& Engineering},
+  volume={9},
+  number={3},
+  year={2007},
+  publisher={IEEE}
+}
+@article{walt2011numpy,
+  title={The NumPy array: a structure for efficient numerical computation},
+  author={Walt, St{\'e}fan van der and Colbert, S Chris and Varoquaux, Gael},
+  journal={Computing in Science \& Engineering},
+  volume={13},
+  number={2},
+  pages={22--30},
+  year={2011},
+  publisher={IEEE}
+}
+@article{oskar2017medpy,
+  title={Medical Image Processing in Python},
+  author={Maier, Oskar},
+  journal={GitHub},
+  volume={},
+  number={},
+  year={2017},
+  publisher={GitHub}
+}
 @article{raamana2014novel,
   title={Novel ThickNet features for the discrimination of amnestic MCI subtypes},
   author={Raamana, Pradeep Reddy and Wen, Wei and Kochan, Nicole A and Brodaty, Henry and Sachdev, Perminder S and Wang, Lei and Beg, Mirza Faisal},
