+++ paper.bib
@@ -46,6 +46,12 @@
   year={2014},
   publisher={Elsevier}
 }
+@techreport{hagberg2008exploring,
+  title={Exploring network structure, dynamics, and function using NetworkX},
+  author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+  year={2008},
+  institution={Los Alamos National Laboratory (LANL)}
+}
 @article{raamana2015thickness,
   title={Thickness network features for prognostic applications in dementia},
   author={Raamana, Pradeep Reddy and Weiner, Michael W and Wang, Lei and Beg, Mirza Faisal and Alzheimer's Disease Neuroimaging Initiative and others},
