### 2017-08-25 0673eaf
badge for code health [skip ci]
