### 2017-08-25 #376 @JohnGriffiths
I'm happy to review this

On 24 August 2017 at 16:06, Christopher Madan <notifications@github.com>
wrote:

> @raamana <https://github.com/raamana> - I'd be happy to serve as editor
> for this submission
>
> @KirstieJane <https://github.com/kirstiejane>, @oesteban
> <https://github.com/oesteban>, @JohnGriffiths
> <https://github.com/johngriffiths> - Are any of you interested/available
> to be a reviewer for this submission? For an overview of the review
> process, please see the reviewer guidelines (http://joss.theoj.org/about#
> reviewer_guidelines)
>
