### 2017-08-25 #376 @cMadan
@arfon, I think I annoyed @whedon... The deleted comment was me trying to start the review and @whedon again not taking the magic-word. :/

@oesteban, @JohnGriffiths - I think it would be great to have both of you as reviewers, though the current JOSS system only allows one person to be 'officially' listed as the reviewer. Regardless, once I get the review started, please both review the submission and make comments accordingly.
