### 2017-08-26 #376 @JohnGriffiths
Ok I can do that.

On 25 August 2017 at 19:58, Pradeep Reddy Raamana <notifications@github.com>
wrote:

> If only one reviewer is needed, perhaps we could use @JohnGriffiths
> <https://github.com/johngriffiths> time to review #379
> <https://github.com/openjournals/joss-reviews/issues/379>? Thanks.
>
