### 2017-08-27 #380 @whedon
[REVIEW]: Histogram-weighted Networks for Feature Extraction, Connectivity and Advanced Analysis in Neuroscience
**Submitting author:** @raamana (<a href="http://orcid.org/0000-0003-4662-0558">Pradeep Reddy Raamana</a>)
**Repository:** <a href="https://github.com/raamana/hiwenet" target ="_blank">https://github.com/raamana/hiwenet</a>
**Version:** 0.4
