### 2017-08-27 #380 @cMadan
@oesteban, it would be great if you could work through the reviewer checklist in the first comment and see if the submission addresses each point sufficiently. If not, please post issues in the submission's repo or here, as makes sense and hopefully @raamana can respond accordingly.

@JohnGriffiths, if you have time, it would still be great if you can also take a look at the submission too (even though you're not officially assigned as a reviewer).
