### 2017-08-28 #376 @raamana
No problem, Kirstie. I was interested in your comments as you are among my target audience :), and wanted to use your comments to improve this package.
