### 2017-08-28 4ad3899
link to guidelines to contribute [skip ci]
