### 2017-08-30 raamana/hiwenet#1 @oesteban
Version in JOSS submission does not match GitHub's latest release
Here (https://github.com/openjournals/joss-reviews/issues/380#issuecomment-326038826), version 0.2.2 is reported but the repo's version indicates 0.2. The installation script specifies 0.2.2 (https://github.com/raamana/hiwenet/blob/master/setup.py#L6).

I would suggest the versioneer package to keep track automatically of versions.

Just generating the appropriate git tag (0.2.2) would satisfy this comment, though.
