### 2017-08-31 raamana/hiwenet#1 @raamana
Thanks Oscar. I'm a little confused here. Are you suggesting me to do this or the JOSS maintainers? :) from what I can see the version number on pypi, on JOSS issue and setup script are all the same.
