### 2017-08-31 raamana/hiwenet#1 @raamana
TBH the recent changes have been super minor - like better packaging etc. One this gets published, version numbers will be much more meaningful with fewer releases of more substantial changes.
