### 2017-08-31 raamana/hiwenet#1 @raamana
I just checked again. From what I can see, the version number on pypi, on JOSS issue and the setup script in the repo here are all the same. Let me know if I am missing something.
