### 2017-08-31 raamana/hiwenet#1 @oesteban
Hi @raamana, I just checked the item in the revision list that says:

> - [ ]  *Version*: Does the release version given match the GitHub release (0.2.2)?

The release version for the submission is 0.2.2 as indicated in that issue, however such version does not exist in your GitHub repo (https://github.com/raamana/hiwenet/releases)

As I said, just  (if your git remotes are standard):

```
git tag 0.2.2
git push origin 0.2.2
```
would suffice to tick the item.

Additionally, I advised not to manually keep track of versions (and let Git do it) by using some software like versioneer. But that is just a hint and hence, not required for acceptance.
