### 2017-08-31 cfd617e
example code for thickness [skip ci]
