### 2017-08-31 83e12ec
concise usage to not overcrowd readme [skip ci]
