### 2017-08-31 raamana/hiwenet#1 @raamana
Gotcha, my bad, I missed the "release" part altogether :) Fixed it. Open this issue again, if I missed something.
