### 2017-08-31 raamana/hiwenet#1 @raamana
thanks for the versioneer suggestion, I will take a look. Manual update is indeed boring and error prone.
