### 2017-08-31 raamana/hiwenet#2 @oesteban
[JOSS Review] Please indicate clearly that the software is not Python 3 compatible
The software only works on Python 2. Given the current size of the project (small) and the fact that Python 2 has been discontinued, I would strongly suggest moving forward and start supporting Python 3 now.

In my opinion, no backwards compatibility is necessary. So it is up to the authors to keep it or dismiss it.

If making this software Python 3 compatible is not in the roadmap, both points should be made very clear in the top README file of the project.

It is true that the authors are using the appropriate classifiers for Pypi, but those classifiers are not taken into account by `pip` during installation. Long story short: installation works in a Python 3 environment without even a warning.

(ref https://github.com/openjournals/joss-reviews/issues/380)
