### 2017-08-31 9d6b4c8
improved example [skip ci]
