### 2017-08-31 raamana/hiwenet#2 @raamana
You're absolutely right - Python 3 comptibility is definitely in the roadmap, just a matter of time for me to work my way through packaging and getting it to work.

From my initial trials, it was not easy (interactions with different packages were not easy to debug). This was suprising as there is nothing in the code that's complicated and not expected to work across multiple versions (atleast PyCharm thinks that way).

I can perhaps take your tips to make it Python 3 compatible.
