### 2017-08-31 43df569
clarification on suport for python 3 [skip ci]
