### 2017-08-31 raamana/hiwenet#2 @oesteban
I would strongly advise to that. I would also advise to drop Python 2 compatibility, even though it may limit (very little, that must be said), the community of users.

But attaching to Python 2 it does really limit the space of users (as a matter of fact I will need to start a python 2 environment which I don't have in my desktop... :D)
