### 2017-08-31 raamana/hiwenet#2 @raamana
updated the readme as suggested.
