### 2017-08-31 raamana/hiwenet#2 @raamana
True, I will move to python 3-only environment in few weeks after my deadlines are done.
