### 2017-08-31 raamana/hiwenet#4 @oesteban
[JOSS Review] Paper.md
I am now checking the last two checkpoints of the review list:

> - [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> - [ ] **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

The statement of need basically replicates the summary in the README file. I think that, for JOSS, a more explicit statement should be made. The target audience is not mentioned in the Summary (of course it can be inferred from the summary and the paper, but I think JOSS wants it to be explicit, not implicit).

Some references from the preprint should be extracted and placed in the references section of Paper.md. At least a reference about the earliest structural connectivity analyses that looked at the cortical thickness (He 2007) and references to the historical methods to extract these networks (there is a table about this in the preprint). It is important to place this contribution in the full scientific scope.

Adding those references will also make it easier to describe the unique contributions of this software.

In my opinion, it is also important to mention that this is a reimplementation of an original code in Matlab. First, I want to applaud the authors' transparency noting this point and their effort of providing a Python version that opens up the audience field. Second, I would advice to include the original Matlab code in the repo, as part of the documentation. That way the authors ensure the transparency of the reimplementation. Third, this point is a bit disturbing to me, and I'm not sure of my opinion on this. Should the Matlab code be the one to be submitted to JOSS?

Additionally, (and this goes beyond this review, it is more for the pre-print), I'd advise to repeat all the analyses done in the paper with this software (if possible), to check that this re-implementation works exactly as the Matlab implementation.

(ref. https://github.com/openjournals/joss-reviews/issues/380)
