### 2017-08-31 raamana/hiwenet#4 @raamana
Thanks Oscar, for going the extra mile of reading the preprint :) and giving me useful suggestions!.

Just to clarify, I didn't intend to connect this to the validity of preprint in any ways - I referred to the preprint as an example to show how it is used in an active research project. In  my mind, there are two separate things - paper would be judged on its scientific and research merit separately. The open sourcing of this reimplementation, as you correctly recognize is to open up this analysis to a broader audience, but also to enable them to add their own new ideas for metrics effortlessly (not just those based on histogram distances). I think the the package needs to be judged on its own merits (not a paper outside its scope). Perhaps I can clarify that in the read me, and that will be enough for the review?
