### 2017-08-31 raamana/hiwenet#4 @raamana
As for verifying this implementation with my older matlab implementation, its a huge task and I expect many differences without even trying, for simple reasons 1) that they are based on different histogram distance toolboxes with different implementations, and 2) this was not aimed to exactly mimic my matlab code bit-by-bit, as I was targeting broader audience (beyong structural & thickness data for example). I did put in a lot of work to write many property tests for verifying metrics in this package are indeed metrics ( check [hiwenet/test_medpy.py](https://github.com/raamana/hiwenet/blob/master/hiwenet/test_medpy.py).

Does that answer your comments?
