### 2017-08-31 7ca670d
clarification on connection to preprint [skip ci]
