### 2017-08-31 raamana/hiwenet#4 @raamana
Updated the readme clarifying the connection to preprint.

PS: reviews like this are useful, making me think about stuff I myself wouldn't have thought about.
