### 2017-08-31 e37bf2d
improved description of target audience [skip ci]
