### 2017-08-31 raamana/hiwenet#4 @raamana
updated the paper.md and README.md as suggested improving the description of target audience
