### 2017-08-31 0922ed0
addition of matlab code to help the matlab users [skip ci]
