+++ matlab_code/README.txt
@@ -1,0 +1,7 @@
+
+These are few source files from my matlab codebase from my original implementation of hiwenet, to give an idea of core computation. 
+
+Please be advised that this code is not meant to work out of the box (like the python package here), just shared to enable matlab users to get started faster. I unfortunately do not have time to maintain matlab code, as I am moving away from it 100%.
+
+Histogram distance implementation is based on the toolbox below:
+http://www.mathworks.com/matlabcentral/fileexchange/39275-histogram-distances
