+++ paper.bib
@@ -52,6 +52,26 @@
   year={2008},
   institution={Los Alamos National Laboratory (LANL)}
 }
+@article{he2007small,
+  title={Small-world anatomical networks in the human brain revealed by cortical thickness from MRI},
+  author={He, Yong and Chen, Zhang J and Evans, Alan C},
+  journal={Cerebral cortex},
+  volume={17},
+  number={10},
+  pages={2407--2419},
+  year={2007},
+  publisher={Oxford University Press}
+}
+@article{lerch2006mapping,
+  title={Mapping anatomical correlations across cerebral cortex (MACACC) using cortical thickness from MRI},
+  author={Lerch, Jason P and Worsley, Keith and Shaw, W Philip and Greenstein, Deanna K and Lenroot, Rhoshel K and Giedd, Jay and Evans, Alan C},
+  journal={Neuroimage},
+  volume={31},
+  number={3},
+  pages={993--1003},
+  year={2006},
+  publisher={Elsevier}
+}
 @article{raamana2015thickness,
   title={Thickness network features for prognostic applications in dementia},
   author={Raamana, Pradeep Reddy and Weiner, Michael W and Wang, Lei and Beg, Mirza Faisal and Alzheimer's Disease Neuroimaging Initiative and others},
