### 2017-08-31 raamana/hiwenet#4 @raamana
Shared the [matlab code here](https://github.com/raamana/hiwenet/tree/master/matlab_code) as suggested.

Added more older references as suggested. Given the nature of a JOSS, I don't think its possible to elaborate on the context and scientific scope, without making paper.md too long, and I am not sure if the JOSS paper is meant to serve that purpose (its more about the package and what it does). I think users can always read the cited references (now including those you suggested) to get the context and scope.

Very helpful comments so far Oscar. That should have covered your comments so far? Let me know if you have any other suggestions.
