### 2017-08-31 raamana/hiwenet#4 @raamana
Also, added the link to matlab code in readme.
