### 2017-08-31 ee9ccba
clarification on matlab code [skip ci]
