### 2017-09-04 94af744
first steps to moving to python 3+
