+++ .travis.yml
@@ -1,7 +1,8 @@
 language: python
 cache: pip
 python:
-  - "2.7"
+  - "3.5"
+  - "3.6"
   
 install:
   - pip install -r requirements.txt
