### 2017-09-04 833fcfe
including 2.7 also until medpy is upgraded to support 3+
