+++ .travis.yml
@@ -1,6 +1,7 @@
 language: python
 cache: pip
 python:
+  - "2.7"
   - "3.5"
   - "3.6"
   
