### 2017-09-04 b0d5c34
changing init behaviour depending on python version
