### 2017-09-04 6aecdd9
refactoring and exporting CLI entry point
