+++ hiwenet/test_hiwenet.py
@@ -9,10 +9,10 @@ from sys import version_info
 
 if version_info.major==2 and version_info.minor==7 and version_info.micro==13:
     from hiwenet import extract as hiwenet
-    from hiwenet import __run as CLI
+    from hiwenet import run_cli as CLI
 elif version_info.major > 2:
     from hiwenet.hiwenet import extract as hiwenet
-    from hiwenet.hiwenet import __run as CLI
+    from hiwenet.hiwenet import run_cli as CLI
 else:
     raise NotImplementedError('hiwenet supports only 2.7.13 or 3+. Upgrate to Python 3+ is recommended.')
 
