+++ hiwenet/__main__.py
@@ -3,7 +3,7 @@ import hiwenet
 def main():
     "Entry point."
 
-    hiwenet.__run()
+    hiwenet.run_cli()
 
 if __name__ == '__main__':
     main()
