### 2017-09-04 raamana/hiwenet#2 @raamana
Hey Oscar, check #6 to see my attempts for compatibility. A dependency is holding me back.
