### 2017-09-05 d7db009
fixing CLI entry point
