### 2017-09-05 raamana/hiwenet#4 @raamana
Hey Oscar, just want to note, the build is failing (as per the badge) because I included python 3+ on CI as well as python 2.7. If you check the sub-builds, you can see it passing tests for 2.7, and failing on python 3+ because of incompatible code in the dependency medpy.
