### 2017-09-05 91532e6
improved python 2.7 and 3.6 support
