+++ hiwenet/test_hiwenet.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 
 from sys import version_info
 
-if version_info.major==2 and version_info.minor==7 and version_info.micro==13:
+if version_info.major==2 and version_info.minor==7:
     from hiwenet import extract as hiwenet
     from hiwenet import run_cli as CLI
 elif version_info.major > 2:
