### 2017-09-06 0e2ed2e
code cleanup and reorg for py3 release
