### 2017-09-06 raamana/hiwenet#7 @oesteban
[JOSS Review] API is not documented
Methods are documented with docstrings, but no automated method to produce the API documentation is available. Using sphinx would be really straightforward to document this project properly.

This software would really improve if documentation was built automatically and served by a third-party service like http://readthedocs.org

(ref. https://github.com/openjournals/joss-reviews/issues/380)
