### 2017-09-06 raamana/hiwenet#8 @oesteban
[JOSS review] Discussion and Clarifications 
From http://joss.theoj.org/about#author_guidelines:

> *What are your requirements for submission?*
>
>  - [x] Your software should be open source as per the OSI definition
>  - [x] Your software should have an obvious research application
>  - [x] You should be a major contributor to the software you are submitting
>  - [x] Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)
>  - [ ] Should be feature complete (no half-baked solutions)
>  - [ ] Minor, 'utility' packages are not acceptable
> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar.

1. The point _Should be feature complete (no half-baked solutions)_ could be ticked on, but some details (as the fact that no automated generation of documentation is provided #7 ) points to the opposite. Tests are probably a bit light as well (will post a different issue with more concrete ideas). These issues can be addressed during the review process.

2. But I'm mostly concerned about the last point: _Minor, 'utility' packages are not acceptable_. The core of the proposed package is [the extract function](https://github.com/raamana/hiwenet/blob/0e2ed2e667c7093d717482ea3dd301fa11421b01/hiwenet/hiwenet.py#L59-L167) which computes the histograms for each ROI in the parcellation and then uses the histogram distance function from medpy to calculate the histogram-weighted network. The code is completed by some eight supporting functions that basically wrap methods from numpy and medpy to address particular operations. I would like to hear from Pradeep the aspects along which this package goes beyond the concept of 'utility'.

    **This is a very valuable piece of code and a great contribution to network analysis**, please do not get me wrong. But I would find it extremely more useful (and accessible to a wider audience) if it was integrated in other neuroimaging packages, like MNE, nilearn, nistats, medpy (which is a dependency) or even nibabel. Was a PR requested to any of these projects and rejected in first place? These packages always credit contributors and reference the original methods, so I don't see the downside.

    Particularly, nilearn currently misses the calculation of histogram-weighted networks and formal utilities for signal extraction from surface (it can be done, but with some manual massaging). Adding these two features (which is, basically what HiWiNet does currently) to nilearn would be extremely useful. And, hiwinets could be calculated on volume-based data as well, at no extra cost.

(ref. https://github.com/openjournals/joss-reviews/issues/380)
