### 2017-09-06 6f58b39
adding DOI for medy [skip ci]
