+++ paper.bib
@@ -28,14 +28,16 @@
   year={2011},
   publisher={IEEE}
 }
-@article{oskar2017medpy,
-  title={Medical Image Processing in Python},
-  author={Maier, Oskar},
-  journal={GitHub},
-  volume={},
-  number={},
-  year={2017},
-  publisher={GitHub}
+@misc{loli_2017_886618,
+  author       = {loli and
+                  Pradeep Reddy Raamana and
+                  mamrehn and
+                  Rémi Bèges},
+  title        = {loli/medpy: MedPy 0.2.2},
+  month        = sep,
+  year         = 2017,
+  doi          = {10.5281/zenodo.886618},
+  url          = {https://doi.org/10.5281/zenodo.886618}
 }
 @article{raamana2014novel,
   title={Novel ThickNet features for the discrimination of amnestic MCI subtypes},
