### 2017-09-06 raamana/hiwenet#8 @raamana
Your points are well taken, Oscar. Here is a very quick repsonse to your points
 - I agree auto-generation of docs is ideal and will work towards it soon.
 - I am curious to learn how my tests were light
 - I think you've answered your own question/critique :). Quickly, I believe in the Unix philosophy for tool building, and the scope of hiwenet was decided with that in mind. Because, the function hiwenet serves is generic (I mean beyond neuroscience, not just a modalities like fMRI or MEG etc), having it separate from the popular aforementioned packages (meaning not restricting it to nilearn or MNE etc) is a strength IMHO. As you know, developers there can easily plug this in (in additon to other types of covariance estimates, as histogram distances is only one family). I imagine (and am hoping to) add more families of distances, as I find time or help from others.
 - Counter example to your comments are (following the Unix philosophy), are two of my favourite commands: `wc` and `split`. They each do only one thing and they do them well. That's what I hope hiwenet is and will be.

Pinging @cmadan for comments.
