+++ paper.bib
@@ -28,8 +28,8 @@
   year={2011},
   publisher={IEEE}
 }
-@misc{loli_2017_886618,
-  author       = {loli and
+@misc{oskar_maier_2017_886618,
+  author       = {Oskar Maier and
                   Pradeep Reddy Raamana and
                   mamrehn and
                   Rémi Bèges},
