### 2017-09-06 raamana/hiwenet#8 @raamana
Also, have you taken a look around on JOSS list of accepted papers?
