### 2017-09-06 raamana/hiwenet#8 @oesteban
> I agree auto-generation of docs is ideal and will work towards it soon.

Nice!

> I am curious to learn how my tests were light

Sure, I'll try to help as much as I can on this :)

> I think you've answered your own question/critique :).

I hope not, my intention is to help. It's not to discuss with myself alone :D

> Quickly, I believe in the Unix philosophy for tool building, and the scope of hiwenet was decided with that in mind. Because, the function hiwenet serves is generic (I mean beyond neuroscience, not just a modalities like fMRI or MEG etc), having it separate from the popular aforementioned packages (meaning not restricting it to nilearn or MNE etc) is a strength IMHO.

I also do believe in the Unix philosophy, but here the discussion is different:

  **is hiwenet a 'utility' or not?**.

As regards the generality of hiwenet, let's be precise (please correct me if I wrongly understood the following). There are two main components:
  - The generation of hiwenets, where the inputs are features and the output is a network
  - Some code around that, to enable taking in surfaces and generating the features table on which the histograms are computed.

The first is the core component of the contribution, and could be used (as you mention) importing it from a different package (strengthening the idea of utility).
The second is necessary for making the tool command-line usable and to enable some more standard outputs (which is fine, but as I mention below, you could use nibabel+numpy or MNE to do the same).

MNE is very generic and provides large support to reading surfaces and applying parcellations to extract signals. This could be brains or whatever object. Examples in nilearn are only on volumes and here you have the example of how to calculate the connectivity (without the histogram weighting which is missing in nilearn): http://nilearn.github.io/connectivity/region_extraction.html#computing-functional-connectivity-matrices. This can be applied to fMRI or whatever 3D data.

> As you know, developers there can easily plug this in (in additon to other types of covariance estimates, as histogram distances is only one family). I imagine (and am hoping to) add more families of distances, as I find time or help from others.

Integrating your code in a well-established package will supercharge the visibility of your contribution. It will also make it more "open-sourcy", not less (as you suggest with the Unix philosophy idea).

> Counter example to your comments are (following the Unix philosophy), are two of my favourite commands: wc and split. They each do only one thing and they do them well. That's what I hope hiwenet is and will be.

They are, as you say, two beautiful _utilities_ that require tests, maintenance, and love in general. Again, this is not the question.

As a reviewer, I just raise the question. Now, the interpretation of my question and your answers belongs in @cMadan's responsibilities as you mentioned.

To wrap-up, as a reviewer, I am trying to push in two directions here:
  - Get an explicit answer to the central question I raised (regardless the interpretation of your answers that corresponds to the editors).
  - Get an idea of how well I described this software above (about the two components of hiwenet). I could be missing something.

In the end, your responses to both points will help better define the Statement of need for this software and the scope of its features.
