### 2017-09-06 raamana/hiwenet#8 @oesteban
> Also, have you taken a look around on JOSS list of accepted papers?

Can you point me to examples of utilities published in JOSS?
