### 2017-09-07 raamana/hiwenet#8 @raamana
Thanks for pointing me to nilearn's connectome measure. Hiwenet will fit in there very well.

Just like the LedoitWolf covariance is imported from sklearn (which they're not re-coding into nilearn or MNE)
https://github.com/nilearn/nilearn/blob/master/nilearn/connectome/connectivity_matrices.py#L8
https://github.com/mne-tools/mne-python/blob/maint/0.14/mne/cov.py#L871

hiwenet could also be integrated very easily, with the sklearn's fit/transform interface.

Yes it is an utility, and I believe its non-trivial, given its obvious research application and scope of utility, which is not minor.

Now that we discuss integration into other softwares, I need to consider implementing the fit/transform interface.
