### 2017-09-07 f4f27a8
improving docstrings for better API reference in rst [skip ci]
