+++ hiwenet/hiwenet.py
@@ -57,23 +57,30 @@ def __compute_bin_edges(features, num_bins, trim_outliers, trim_percentile):
     return edges
 
 
-def extract(features, groups, weight_method=default_weight_method,
-            num_bins=default_num_bins, trim_outliers=True, trim_percentile=default_trim_percentile,
+def extract(features, groups,
+            weight_method=default_weight_method,
+            num_bins=default_num_bins,
+            trim_outliers=True, trim_percentile=default_trim_percentile,
             return_networkx_graph=False, out_weights_path=None):
     """
-    Extracts the histogram weighted network.
+    Extracts the histogram-distance weighted adjacency matrix.
     
     Parameters
     ----------
-    features : numpy 1d array of length p
-        with scalar values.
+    features : numpy 1d array
+        1d array of scalar values
+
     groups : numpy 1d array
-        Membership array of length p, each value specifying which group that particular node belongs to.
-        For example, if you have a cortical thickness values for 1000 vertices belonging to 100 patches,
-        this array could  have numbers 1 to 100 specifying which vertex belongs to which cortical patch.
+        Membership array of same length as `features`, each value specifying which group that particular node belongs to.
+
+        For example, if you have cortical thickness values for 1000 vertices (`features` is ndarray of length 1000),
+        belonging to 100 patches, the groups array (of length 1000) could  have numbers 1 to 100 (number of unique values)
+        specifying which element belongs to which cortical patch.
+
         Grouping with numerical values (contiguous from 1 to num_patches) is strongly recommended for simplicity,
-        but this could also be a list of strings of length p, in which case a tuple is
-        returned identifying which weight belongs to which pair of patches.
+        but this could also be a list of strings of length p, in which case a tuple is returned,
+        identifying which weight belongs to which pair of patches.
+
     weight_method : string, optional
         Type of distance (or metric) to compute between the pair of histograms.
 
@@ -104,7 +111,8 @@ def extract(features, groups, weight_method=default_weight_method,
         - 'relative_bin_deviation'
         - 'relative_deviation'
 
-        Note only the following are metrics:
+        Note only the following are *metrics*:
+
         - 'manhattan'
         - 'minowski'
         - 'euclidean'
@@ -112,7 +120,8 @@ def extract(features, groups, weight_method=default_weight_method,
         - 'noelle_4'
         - 'noelle_5'
 
-        The following are semi- or quasi-metrics:
+        The following are *semi- or quasi-metrics*:
+
         - 'kullback_leibler'
         - 'jensen_shannon'
         - 'chi_square'
@@ -127,6 +136,7 @@ def extract(features, groups, weight_method=default_weight_method,
         - 'noelle_3'
 
         The following are  classified to be similarity functions:
+
         - 'histogram_intersection'
         - 'correlate'
         - 'cosine'
@@ -138,31 +148,45 @@ def extract(features, groups, weight_method=default_weight_method,
 
     num_bins : scalar, optional
         Number of bins to use when computing histogram within each patch/group.
+
         Note:
+
         1) Please ensure same number of bins are used across different subjects
-        2) histogram shape can vary widely with number of bins (esp with fewer bins in the range of 3-20),
-        and hence the features extracted based on them vary also.
+        2) histogram shape can vary widely with number of bins (esp with fewer bins in the range of 3-20), and hence the features extracted based on them vary also.
         3) It is recommended to study the impact of this parameter on the final results of the experiment.
+
         This could also be optimized within an inner cross-validation loop if desired.
+
     trim_outliers : bool, optional
-        Whether to trim 5% outliers at the edges of feature range,
[…]
