+++ docs/API.rst
@@ -1,0 +1,20 @@
+
+API Reference
+--------------
+
+A tutorial-like presentation is available at :doc:`usage`, using the following API.
+
+.. automodule:: hiwenet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Command line interface
+-----------------------
+
+.. argparse::
+   :ref: hiwenet.__get_parser
+   :prog: hiwenet
+   :nodefault:
+   :nodefaultconst:
+
