+++ docs/metrics.rst
@@ -1,0 +1,5 @@
+Metrics
+-------
+
+Definition of the different distances between the histograms are presented `here <http://pythonhosted.org//MedPy/metric.html#module-medpy.metric.histogram>`_.
+
