+++ docs/installation.rst
@@ -1,0 +1,16 @@
+------------
+Installation
+------------
+
+This package could easily be installed via:
+
+``pip install hiwenet``
+
+
+Requirements
+------------
+
+ - numpy
+ - medpy
+ - hypothesis
+ - networkx
