### 2017-09-07 6844e07
ReadTheDocs mocking out numpy stuff
