### 2017-09-07 a742b1b
ReadTheDocs - adding numpy and sphinx-argparse as requirements
