+++ docs/conf.py
@@ -48,6 +48,8 @@ extensions = ['sphinx.ext.autodoc', 'numpydoc', 'sphinxarg.ext',
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages']
 
+requirements_file: rtdocs_requirements.txt
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
