+++ docs/rtdocs_requirements.txt
@@ -1,0 +1,2 @@
+numpy
+sphinx-argparse
