### 2017-09-07 929669a
fixing conf.py syntax error [skip ci]
