+++ docs/conf.py
@@ -48,7 +48,7 @@ extensions = ['sphinx.ext.autodoc', 'numpydoc', 'sphinxarg.ext',
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages']
 
-requirements_file: rtdocs_requirements.txt
+requirements_file = 'rtdocs_requirements.txt'
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
