### 2017-09-07 46fe93d
another try to fix conf.py syntax error [skip ci]
