+++ docs/rtdocs_requirements.txt
@@ -1,2 +1,3 @@
 numpy
+numpydoc
 sphinx-argparse
