### 2017-09-07 20fdbc8
final try to fix conf.py syntax error [skip ci]
