+++ docs/rtdocs_requirements.txt
@@ -1,3 +1,4 @@
 numpy
 numpydoc
 sphinx-argparse
+networkx
