### 2017-09-07 d6034f5
keep trying to fix conf.py syntax error [skip ci]
