### 2017-09-07 50d568d
Revert "keep trying to fix conf.py syntax error [skip ci]"
This reverts commit d6034f5
