+++ docs/rtdocs_requirements.txt
@@ -2,8 +2,3 @@ numpy
 numpydoc
 sphinx-argparse
 networkx
-medpy
-hypothesis
-networkx
-matplotlib
-codecov
