### 2017-09-07 cb655ae
keep trying to fix conf.py syntax error [skip ci]
