### 2017-09-07 raamana/hiwenet#7 @raamana
Hey Oscar, check the docs website now: http://hiwenet.readthedocs.io

The API page compiles properly locally, but not on RTD. It did once, it seems it is flaky when using sphinx-argparse extension. Will work with them to fix it.
