### 2017-09-07 raamana/hiwenet#7 @raamana
Actually, it just worked again :). Open this if you have suggestions to improve.

Thanks for your comments Oscar, really improving the repo.
