### 2017-09-07 c4fc23b
Adding references and guidelines [skip ci]
