+++ docs/metrics.rst
@@ -3,3 +3,5 @@ Metrics
 
 Definition of the different distances between the histograms are presented `here <http://pythonhosted.org//MedPy/metric.html#module-medpy.metric.histogram>`_.
 
+
+**I plan to add more families of distances and metrics soon, your comments on what could be good additions will be appreciated, or links to similar repositories that I can link to will be appreciated.**
