### 2017-09-07 raamana/hiwenet#4 @raamana
Do my comments answer your questions in this issue Oscar? Let me know if otherwise.
