### 2017-09-07 raamana/hiwenet#7 @oesteban
Thanks a lot, this is a lot more convenient.

Some minor suggestions:

- Usage section should be renamed to Examples
- The command line interface section should not be under the API section. You could extract it and name it Usage (this time it is appropriate). One example of command line (one line) would be very much appreciated.
- Add a section "original implementation" where you include the original Matlab code and comment on which ways this implementation differs.
