### 2017-09-07 raamana/hiwenet#4 @oesteban
Sorry for the wait on this issue.

> Just to clarify, I didn't intend to connect this to the validity of preprint in any ways - I referred to the preprint as an example to show how it is used in an active research project. In my mind, there are two separate things - paper would be judged on its scientific and research merit separately. The open sourcing of this reimplementation, as you correctly recognize is to open up this analysis to a broader audience, but also to enable them to add their own new ideas for metrics effortlessly (not just those based on histogram distances). I think the the package needs to be judged on its own merits (not a paper outside its scope). Perhaps I can clarify that in the read me, and that will be enough for the review?

By linking the pre-print you are implicitly making an statement of need: "this software allows you to do this, as shown in [your preprint]". The main figure is shared in both, which also points in that direction. Your claim "_there are two separate things - paper would be judged on its scientific and research merit separately_" is correct. Therefore, how preprint and code relate to one another should be explicitly stated whenever you use the preprint to support the software: "This is a re-implementation of [your method], based on the experience of [your preprint]", indicating that "results obtained with this library may differ from what you'd get with the original Matlab code". Also, since you have made your Matlab code public, adding the pointer to the original implementation will be really nice: "Here, you will find the original Matlab code we used in the paper just in case you want to replicate our findings" (a documentation page with the script would be great, I added a comment on this to #7). I presume that you will cite this JOSS publication in your preprint so I totally agree with your view (the two things address different facets of scientific communication) but you need to say clearly how and why they differ (IMHO). Otherwise, one could try to reproduce your results without knowing that they are using a different implementation. All of this is (again IMHO) particularly important for the JOSS' Paper.md.

This leads me to the following point.

> updated the paper.md and README.md as suggested improving the description of target audience

I think you that the README (which it is probably good enough as it is now, I'll confirm this point when the review is finished) and the Paper.md should be dissociated as they are very different entities with different purposes (please, reformulate Paper.md).

> Added more older references as suggested. Given the nature of a JOSS paper, I don't think it is possible to elaborate on the context and scientific scope, without making paper.md too long, and I am not sure if the JOSS paper is meant to serve that purpose (its more about the package and what it does). I think users can always read the cited references (now including those you suggested) to get the context and scope.

(could you point me to the added references?, sorry I don't see them)

As I introduced before, I think the Paper.md should flow as follows:

1. The box you already have on top with authors and affiliation

2. Statement of need, including:

    2.1. Problem: people want to calculate connectivity matrices based on correlations of structural and functional data (with a reference to support this, probably He 2007?). [2-3 lines]

    2.2. Say the purpose of HiWiNet trying to be exact and precise (IMHO you should detach from the definitions in your preprint and think about what exactly your submission does). Cite your preprint here as the main example of application. Indicate the existence of an original Matlab code used in that reference under the documentation folder. The figure may be useful here (I wasn't proposing to remove it before) [4-5 lines].

    2.3. Say why HiWiNet is different (core statement of need) to existing software. I gave some examples like MNE (http://martinos.org/mne/stable/generated/mne.SourceEstimate.html#mne.SourceEstimate.extract_label_time_course) or nilearn (http://nilearn.github.io/connectivity/functional_connectomes.html#extracting-signals-on-a-parcellation). Reference those projects.

    [5-10 lines].

    I think the following addition goes in the right direction, but a clear sentence describing the functionality of HiWiNet with accuracy is still missing.

    > The target audience is users of almost all neuroimaging modalities who:
    > 1. preprocessed dataset already,
    > 2. have some base features extracted (node- or patch-wise, that are native to the given modality) using other packages (metioned above), and
    > 3. who would like to analyze network-level (i.e. covariance-type or connectivity) relations among the base features (either in space across the cortex or a relevant domain, or across time).
    > 4. This is similar to popular metrics of covariance like Correlation or LedoitWolf, and could be dropped in their place. Do you want to find out how histogram-based method compare to your own ideas?

3. References: the 4 you have so far (one from 2.1, your preprint from 2.2, and the existing software of 2.3).
[8 lines]

Total: 26 lines, not too long.

> Hey Oscar, just want to note, the build is failing (as per the badge) because I included python 3+ on CI as well as python 2.7. If you check the sub-builds, you can see it passing tests for 2.7, and failing on python 3+ because of incompatible code in the dependency medpy.

Let's follow up that discussion in #6.
