### 2017-09-07 raamana/hiwenet#8 @oesteban
> I need to consider implementing the fit/transform interface.
This would be a great addition.

> Yes it is an utility, and I believe its non-trivial, given its obvious research application and broad scope of utility, which is not minor.

This touches upon the central question discussed. However, I feel that the minor/major adjective should be analyzed in the space of software engineering and the context of existing software. Scientifically, I have no doubt this is a major contribution necessary to build your preprint and future papers, but IMHO you want to justify how this utility is also a major addition to the neuroimaging-in-python ecosystem.

The following questions are still open:

  - [ ] Get an idea of how well I described this software above (about the two components of hiwenet). I could be missing something - With the addition of #7 I am not sure now that the second component I mentioned up in https://github.com/raamana/hiwenet/issues/8#issuecomment-327639942 does even exist in hiwinet.
  - [ ] https://github.com/raamana/hiwenet/issues/8#issuecomment-327641795 - The answer to this question will be highly valuable to @cMadan and the JOSS editors. If you find some "minor utilities" already published in JOSS, all this discussion would be irrelevant and your submission accepted without doubts. Then, JOSS editors should work on better define that requirement with an exact definition of what is major and minor, and the coordinates in which that should be evaluated.
