### 2017-09-07 raamana/hiwenet#10 @oesteban
[JOSS Review] Software testing
The main missing point here is a "smoke test" (just checking that the software worked, allowing any size of error at the output). How it would work:

1. Generate an example features file. Store it in your repo so travis fetches it.
2. Generate (locally) the outputs from that specific features file. Add it to the repo.
3. Run the command line on that file (use coverage if you want to make sure that run is taken into account for coverage measure, allowing you to get rid of some tests -I go deeper on this below-).
4. Check that the output of 3) is the same you got in 2).

This does not check that the method worked correctly, it only checks that it did not break down (no smoke came out the box).

But, **you happen to have an alternative implementation**. This is a great _testing oracle_. You will be able to 1) ensure your implementation matches the original; 2) be more certain about the correctness of your software. To do so, modify 2) in the previous list with the outputs generated with your Matlab implementation. In 4) you probably want to modify your test to accept certain threshold of numerical tolerance. After all, *you wouldn't need to repeat all the experiments of your preprint to provide some evidence that this implementation is equivalent*.

The rest of the tests just check the conformity of inputs and outputs (https://github.com/raamana/hiwenet/blob/c4fc23b837f072888f5a359ac4df0e1c38ba5d8e/hiwenet/test_hiwenet.py#L46-L110), which is fine.

Or they test argparse, when it already takes care of the correctness and interpretation of the command line. I think these tests do not exercise much the code: https://github.com/raamana/hiwenet/blob/c4fc23b837f072888f5a359ac4df0e1c38ba5d8e/hiwenet/test_hiwenet.py#L113-L168 .

(ref. https://github.com/openjournals/joss-reviews/issues/380)
