### 2017-09-07 raamana/hiwenet#7 @raamana
Done.

I don't think the reference to original implementation is so important.
