### 2017-09-07 #380 @oesteban
# Review comments

## General

This submission describes a software tool to calculate histogram weighted networks (HiWiNet). As far as I understood the code, the core of the tool is the `extract` function which takes an array of features and generates the corresponding network. To do so, the software strongly relies on the histogram distance implementation of `medpy`. The submission includes a command line interface so that the users can run the software directly on a shell prompt with text files for inputs and outputs.

In my opinion, this submission can be accepted after a **Major revision** responding to the points raised in the issues enlisted above. Particularly, the editors of JOSS may want to track https://github.com/raamana/hiwenet/issues/8 closely, as I think we hit there an area of JOSS' requirements for submissions which is (at the very least) gray and confusing - just a spoiler: _what is a "minor" 'utility'_?, _what is exactly a utility_?, and _when a utility is a minor contribution_?.

## Open questions
  - [ ] The `Paper.md` and the `README` of this project are twins, and I think they should be dissociated. Particularly, because for now there is no clear definition of the functionality of the software, and all descriptions derive from a corresponding pre-print. The statement of need will require some maturation and work to make it explicit and accurate.
  - [ ] Implementing a real "smoke test" is at reach, and I think necessary for this project.

## Items that need work (as per the checklist above)

> ### Functionality
>
>  - [ ] **Functionality:** Have the functional claims of the software been confirmed?
>  - [ ] **Performance:** Have any performance claims of the software been confirmed?

- A more precise definition of the functionality should be done to evaluate this point.
- No performance claims are done, so it would be nice to include some. Particularly interesting is the cross-comparison to the former implementation in Matlab that I suggested.

> ### Documentation
>
>  - [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>  - [ ] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

The statement of need is a bit unclear. I think the automated tests need a "smoke test" at the least.

> ### Software paper
>  - [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>  - [ ] **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

The `Paper.md` needs to be made clearer, more exact and follow closer the guidelines, instead of just mirroring of the `README` of the project..
