### 2017-09-07 raamana/hiwenet#7 @oesteban
The fact that there is a primitive/original implementation in Matlab is irrelevant, just a circumstance. However, being transparent about the fact is very important.
