### 2017-09-07 raamana/hiwenet#8 @raamana
Oscar and Chris, I could certainly give you examples of JOSS accepted software, which I would consider minor utilities (download scripts, calculator of few scores, or applicators of few simple steps, visualization scripts etc), if I were to apply your reasoning (which if I understand correctly is only few functions, doing only one thing, not sufficiently distinguished from existing tools etc).

I am choosing to not name them (and leave it upto you and JOSS editors to develop your own opinion: http://joss.theoj.org/papers/popular?page=14), out of a simple reason, that it would be disrespectful to call them minor utilities even if they serve a handful of people well (regardless of the size of the task). This is a personal opinion and its okay if we disagree - but I personally recognize that most of these are open source, have been made in the spirit of open science and they have already volunteered their time and skills to develop, maintain and share them. I believe you also share many of these ideas.

If I were a JOSS reviewer myself, I wouldn't support duplication of functionality, misappropriation of credit, code stealing and the other trivial things etc.  I would judge the utility of a submission by the quality of product delivered (in the mind of the end-user -- not the reviewer)
 - will it improve the end-user's productivity (again, not the reviewers)? Even if its a small section of people, if it is helpful beyond the trivial, it is worth being considered.
 - is it helpful to have a common utility that is well done, tested and reliable, than each user implementing their own (even if it only takes a week to implement it)?

Thanks to your expert review and generous volunteering of time, its been quite improved, and I will continue to improve this as I am able to (given I use it for my research projects). That said, this is still a side project for me, and I will focus on using it for now and let the users drive the development. I may have to reconsider whether to put time towards its publication (not as enthusiastic anymore) or its development as I see it will be helpful to myself and others.
