### 2017-09-07 8570dce
reorg of docs [skip ci]
