+++ docs/API.rst
@@ -9,12 +9,4 @@ A tutorial-like presentation is available at :doc:`usage`, using the following A
    :undoc-members:
    :show-inheritance:
 
-Command line interface
------------------------
-
-.. argparse::
-   :ref: hiwenet.__get_parser
-   :prog: hiwenet
-   :nodefault:
-   :nodefaultconst:
 
