### 2017-09-07 e100756
adding acknowledgements [skip ci]
