### 2017-09-07 raamana/hiwenet#8 @raamana
You were incredibly helpful in all the ways -- your expert review and most importantly your time and consideration will be appreciated by me. In fact, I've just added acknowledgements towards that now..
https://github.com/raamana/hiwenet/blob/master/docs/about.rst#acknowledgements

which should be updated on its website shortly.
