### 2017-09-08 raamana/hiwenet#8 @cMadan
Thank you for the insightful discussion re: JOSS' inclusion criteria. As editors we have also had discussions regarding the scope of the journal and if specific submissions are sufficiently substantial in function (e.g., minor utility, research application). I will direct the JOSS editors to this thread to further our discussions in how we can make the journal criteria clearer.

I think the current submission is sufficiently substantive, but this discussion does nonetheless signal that the function of the submission should be made clearer. In this respect, the statement of need (in the `paper.md`) should be more explicit and demonstrative. It currently gives a broad direction of function, but the addition of a specific statement (only 1 paragraph or so) and related demo dataset would make the submission unarguably sufficient for JOSS. Along these lines, some text about the existing tools that the submission would complement (as discussed, nilearn and mne-tools) would make the submission more approachable to potential users.

I would also strongly suggest that there be additional references in the paper, to provide better context of the submission. I'm sure there are another 2-3 references beyond the listed preprint that can help convey to a reader what the toolbox does. As it is there is only one paragraph that somewhat conveys the functionality, which I myself find a bit opaque. The paper is intended to not be overly burden-some to the developer, but it also needs to describe the submission's function to a slightly broader audience than those who may use it. (I also think that words should be written out fully, "esp.".)
