### 2017-09-08 #380 @cMadan
Thank you for the detailed comments, @oesteban! I have provided some additional (and convergent) comments here https://github.com/raamana/hiwenet/issues/8#issuecomment-328100590.
