### 2017-09-08 f04de21
more details on function [skip ci]
