+++ docs/API.rst
@@ -2,7 +2,7 @@
 API Reference
 --------------
 
-A tutorial-like presentation is available at :doc:`usage`, using the following API.
+A tutorial-like presentation is available at :doc:`usage_api`, using the following API.
 
 .. automodule:: hiwenet
    :members:
