### 2017-09-08 035e43a
allowing file input for features and groups even via API
