+++ docs/paper.bib
@@ -9,6 +9,43 @@
 	eprint = {http://www.biorxiv.org/content/early/2017/07/31/170381.full.pdf},
 	journal = {bioRxiv}
 }
+@article{tijms2012similarity,
+  title={Similarity-based extraction of individual networks from gray matter MRI scans},
+  author={Tijms, Betty M and Seri{\`e}s, Peggy and Willshaw, David J and Lawrie, Stephen M},
+  journal={Cerebral Cortex},
+  volume={22},
+  number={7},
+  pages={1530--1541},
+  year={2012},
+  publisher={Oxford University Press}
+}
+@article{palaniyappan2015abnormalities,
+  title={Abnormalities in structural covariance of cortical gyrification in schizophrenia},
+  author={Palaniyappan, Lena and Park, Bert and Balain, Vijender and Dangi, Raj and Liddle, Peter},
+  journal={Brain Structure and Function},
+  volume={220},
+  number={4},
+  pages={2059--2071},
+  year={2015},
+  publisher={Springer}
+}
+@article{xu2017abnormalities,
+  title={Abnormalities in structural covariance of cortical gyrification in Parkinson's disease},
+  author={Xu, Jinping and Zhang, Jiuquan and Zhang, Jinlei and Wang, Yue and Zhang, Yanling and Wang, Jian and Li, Guanglin and Hu, Qingmao and Zhang, Yuanchao},
+  journal={Frontiers in Neuroanatomy},
+  volume={11},
+  year={2017},
+  publisher={Frontiers Media SA}
+}
+@article{evans2013networks,
+  title={Networks of anatomical covariance},
+  author={Evans, Alan C},
+  journal={Neuroimage},
+  volume={80},
+  pages={489--504},
+  year={2013},
+  publisher={Elsevier}
+}
 @article{oliphant2007python,
   title={Python for scientific computing},
   author={Oliphant, Travis E},
