### 2017-09-08 058c902
adding necessary files for docs generation [skip ci]
