+++ docs/guidelines.rst
@@ -1,0 +1,19 @@
+----------
+Guidelines
+----------
+
+Here I will try to collect relevant tips and references from existing research.
+
+References
+----------
+
+ - `Why is Euclidean distance not a good metric in high dimensions? <https://stats.stackexchange.com/questions/99171/why-is-euclidean-distance-not-a-good-metric-in-high-dimensions>`_
+ - `On the Surprising Behavior of Distance Metrics in High Dimensional Space <https://bib.dbvis.de/uploadedFiles/155.pdf>`_
+ - `Definition of Metrics <https://en.wikipedia.org/wiki/Metric_(mathematics)>`_
+ - `Distance metric learning: A survey <https://s3.amazonaws.com/academia.edu.documents/36879449/dist-metric-survey.pdf?AWSAccessKeyId=AKIAIWOWYYGZ2Y53UL3A&Expires=1504793624&Signature=%2FXy8dcP%2Bsw62a0CR6OZLkmn5ZBk%3D&response-content-disposition=inline%3B%20filename%3DDistance_Metric_Learning_A_Comprehensive.pdf>`_
+
+Neuroscience
+------------
+
+- Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381. doi: https://doi.org/10.1101/170381
+- Others to come - please make a PR to add your papers and others you know of.
