+++ docs/guidelines.rst
@@ -15,5 +15,9 @@ References
 Neuroscience
 ------------
 
-- Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381. doi: https://doi.org/10.1101/170381
-- Others to come - please make a PR to add your papers and others you know of.
+ - Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381
+ - Palaniyappan, L., Park, B., Balain, V., Dangi, R., & Liddle, P. (2015). Abnormalities in structural covariance of cortical gyrification in schizophrenia. Brain Structure and Function, 220(4), 2059-2071.
+ - Xu, J., Zhang, J., Zhang, J., Wang, Y., Zhang, Y., Wang, J., ... & Zhang, Y. (2017). Abnormalities in structural covariance of cortical gyrification in Parkinson's disease. Frontiers in Neuroanatomy, 11.
+ - Raamana, P. R., Wen, W., Kochan, N. A., Brodaty, H., Sachdev, P. S., Wang, L., & Beg, M. F. (2014). Novel ThickNet features for the discrimination of amnestic MCI subtypes. NeuroImage: Clinical, 6, 284-295.
+ - Tijms, B. M., Seriès, P., Willshaw, D. J., & Lawrie, S. M. (2012). Similarity-based extraction of individual networks from gray matter MRI scans. Cerebral Cortex, 22(7), 1530-1541.
+ - Others to come - please make a PR to add your papers and others you know of.
