### 2017-09-08 45fe0bc
better text / typos [skip ci]
