+++ docs/paper.md
@@ -25,14 +25,14 @@ bibliography: paper.bib
 
 # Summary
 
-Network-level analysis of various features, esp. if it can be individualized for a single-subject, is proving to be quite a valuable tool in many applications. This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node. Individual nodes could be an ROI or a patch or a cube, or any other unit of relevance in your application. This is a great way to take advantage of the full distribution of values available within each node, relative to the simpler use of averages (or another summary statistic). 
+Network-level analysis of various features, especially if it can be individualized for a single-subject, is proving to be a valuable tool in many applications. This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node. Individual nodes could be an ROI or a patch or a cube, or any other unit of relevance in your application. This is a great way to take advantage of the full distribution of values available within each node, relative to the simpler use of averages (or another summary statistic). 
 
 Rough scheme of computation is shown below:
 ![illustration](illustration.png)
 
 ## Note on applicability and target audience
 
-Although this technique was originally developed for cortical thickness, this is a generic and powerful technique that could be applied to any features such as gray matter density, PET uptake values, functional activation data or EEG features. All that is needed is a set of nodes/parcellation that have one-to-one correspondence across samples/subjects in your dataset.
+Although this technique was originally developed for cortical thickness analysis in neuroimaging research, this is a generic and powerful technique that could be applied to any features such as gray matter density, PET uptake values, functional activation data, EEG features or any other domain. All that is needed is a set of nodes/parcellation that have one-to-one correspondence across samples/subjects in your dataset.
 
 The target audience is users of almost all neuroimaging modalities who 1) preprocessed dataset already, 2) have some base features (node- or patch-wise, that are native to the given modality) extracted using other packages (metioned above), and 3) who would like to analyze network-level or covariance-type or connectivity relations among the base features.
 
