### 2017-09-08 raamana/hiwenet#8 @raamana
Thanks Chris and Oscar for the clarification.

I didn't mean the paper.md to be static (as it was at the submission time) - it was prepared to give an idea of what I had mind (with the intention to improve it and not leave it as is), for examples with your comments here. Anyways, I've added another para now (what does hiwenet do?), with references to the API/CLI docs, with concrete examples (and relevant files for a demo dataset).

Just to push back on the lack of clarity a bit :), the very second sentence in the first paragraph of paper.md says "**This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node**" -- I am not sure if that can be made any more specific, without code/data, both of which are visible in the repo and examples folder. But, I understand we should try do our best to clarify the purpose and functionality in as many ways as possible, including having to repeat it multiple times at multiple places.

I've mentioned the other tools also and how hiwenet relates to them. As I am planning to add a scikit-learn compatible API, I will include more specific examples directly related to nilearn and MNE - but this will happen once I finalize few more things -- as I would like to survey the landscape even more on what interfaces would be great and what tools to target etc.

docs are here: http://hiwenet.readthedocs.io/en/latest/ - I will try to make that the go-to place for documentation (probably even removing README.md in repo to avoid overlap with paper.md, if needed).

Perhaps it was not clear, but paper.bib had already contained more than my preprint. I've added few more. I am sure you agree with me the function of paper.bib is not meant to be lit survey (Author guidelines need to expanded substantially on what is expected).

I am glad this review is helping to improve the process of JOSS review. Thanks for your time to improve hiwenet. Much appreciate it.
