### 2017-09-08 raamana/hiwenet#8 @oesteban
> This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node

Very accurate 👍

I just wanted to remind my request of a smoke test in travis. The idea shouldn't take more than 20 minutes to implement, and I'm sure Pradeep has a features file ready to use for that test.

Thanks Pradeep for the effort and for a great contribution.
