### 2017-09-08 raamana/hiwenet#10 @raamana
Man, you're incredibly patient to write a detailed issue 👍

Anyways if I understand you and suggested test correctly, this will be testing the I/O aspects of hiwenet, more so than the implementation itself. Yes? I think this is important if we were to offer both interfaces -- working on it.
