### 2017-09-08 raamana/hiwenet#10 @oesteban
I proposed two "flavors":

1. Vanilla:
  - exercises the i/o aspects (as you will see, the coverage will not drop if you remove all the tests on the command line).
  - but more importantly: will let you know when something is broken inside the extract function (meaning, you make a mistake coding and you catch the problem early)

2. Cross-validated:
  - in addition to the vanilla, you will provide proof that the python implementation is close enough to the matlab implementation
