### 2017-09-08 9bcc307
ensuring API and CLI results match, tightly controlling I/O; tests
