### 2017-09-08 raamana/hiwenet#10 @raamana
The 9bcc307dbcd28c96b739eeaa989b3264fd2eef49 should resolve item 1 here (involving API, CLI an I/O).

The second one is more about matching [medpy's](GitHub.com/loli/medpy) implementation with that coming from the [matlab toolbox](http://www.mathworks.com/matlabcentral/fileexchange/39275-histogram-distances). This comparison is complicated and non-trivial, which I feel is slightly beyond the scope of this package/review. It is important for my research and I will dig into it as I try to make comparisons between them.
