### 2017-09-08 c154abf
concrete links to nilearn and MNI [skip ci]
