### 2017-09-08 b2400d6
direct references to nilearn and MNE
