### 2017-09-08 raamana/hiwenet#4 @raamana
I think the last few commits have addressed these suggestions. I esp. like your suggestion 2.3, added it to paper.md and usage_api.rst. Let me know if you suggest anything else.
