### 2017-09-11 raamana/hiwenet#8 @cMadan
> I didn't mean the paper.md to be static (as it was at the submission time) - it was prepared to give an idea of what I had mind (with the intention to improve it and not leave it as is), for examples with your comments here. Anyways, I've added another para now (what does hiwenet do?), with references to the API/CLI docs, with concrete examples (and relevant files for a demo dataset).

I'm not sure I follow... I feel like the project at the time of submission is intended to be 'ready to be accepted' from the submitter's perspective. So, the `paper.md` should have been complete at that time, not that the reviewer and editor point out how it was incomplete and the author says 'it wasn't supposed to be reviewed yet, it was just a placeholder'. Regardless, the new "What does the hiwenet package do?" section does much to address my prior concerns.

> Perhaps it was not clear, but paper.bib had already contained more than my preprint. I've added few more. I am sure you agree with me the function of paper.bib is not meant to be lit survey (Author guidelines need to expanded substantially on what is expected).

The paper.bib is meant to be the references to go along with the paper.md. It should only have references cited in the paper.md. Reciprocally, all entries in paper.bib should be cited in paper.md--I will let you determine how you wish to make these two files align. This is definitely not intended to be a lit review, but also no one will look at the paper.bib file directly, it's only necessary for paper.md to be compiled into the resulting pdf, and as it is, paper.md only has a single reference--your preprint.
