### 2017-09-17 5540b3f
relaxing version check [skip ci]
