### 2017-09-18 e54e1ab
new options to fix edge range (to allow same range across multiple calls)
