+++ hiwenet/hiwenet.py
@@ -3,6 +3,7 @@ __all__ = ['extract', 'run_cli']
 import argparse
 import os
 import sys
+import collections
 import warnings
 import logging
 import networkx as nx
@@ -39,20 +40,24 @@ minimum_num_bins = 5
 
 default_weight_method = 'manhattan'
 default_num_bins = 25
+default_edge_range = None
 default_trim_percentile = 5
 default_trim_behaviour = True
 default_return_networkx_graph = False
 default_out_weights_path = None
 
-def __compute_bin_edges(features, num_bins, trim_outliers, trim_percentile):
+def __compute_bin_edges(features, num_bins, edge_range, trim_outliers, trim_percentile):
     "Compute the edges for the histogram bins to keep it the same for all nodes."
 
-    if trim_outliers:
-        # percentiles_to_keep = [ trim_percentile, 1.0-trim_percentile] # [0.05, 0.95]
-        edges_of_edges = np.array([np.percentile(features, trim_percentile),
-                                   np.percentile(features, 100 - trim_percentile)])
+    if edge_range is None:
+        if trim_outliers:
+            # percentiles_to_keep = [ trim_percentile, 1.0-trim_percentile] # [0.05, 0.95]
+            edges_of_edges = np.array([np.percentile(features, trim_percentile),
+                                       np.percentile(features, 100 - trim_percentile)])
+        else:
+            edges_of_edges = np.array([np.min(features), np.max(features)])
     else:
-        edges_of_edges = np.array([np.min(features), np.max(features)])
+        edges_of_edges = edge_range
 
     # Edges computed using data from all nodes, in order to establish correspondence
     edges = np.linspace(edges_of_edges[0], edges_of_edges[1], num=num_bins, endpoint=True)
@@ -63,6 +68,7 @@ def __compute_bin_edges(features, num_bins, trim_outliers, trim_percentile):
 def extract(features, groups,
             weight_method=default_weight_method,
             num_bins=default_num_bins,
+            edge_range=default_edge_range,
             trim_outliers=default_trim_behaviour,
             trim_percentile=default_trim_percentile,
             return_networkx_graph=default_return_networkx_graph,
@@ -164,6 +170,20 @@ def extract(features, groups,
 
         This could also be optimized within an inner cross-validation loop if desired.
 
+    edge_range : tuple or None
+        The range of edges within which to bin the given values.
+        This can be helpful to ensure correspondence across multiple invocations of hiwenet (for different subjects),
+        in terms of range across all bins as well as individual bin edges.
+        Default is to automatically compute from the given values.
+
+        Accepted format:
+
+            - tuple of finite values: (range_min, range_max)
+            - None, triggering automatic calculation (default)
+
+        Notes : when controlling the ``edge_range``, it is not possible trim the tails (e.g. using the parameters
+        ``trim_outliers`` and ``trim_percentile``) for the current set of features using its own range.
+
     trim_outliers : bool, optional
         Whether to trim a small percentile of outliers at the edges of feature range,
         when features are expected to contain extreme outliers (like 0 or eps or Inf).
@@ -177,7 +197,8 @@ def extract(features, groups,
         Specifies the need for a networkx graph populated with weights computed. Default: False.
 
     out_weights_path : str, optional
-        Where to save the extracted weight matrix. If networkx output is returned, it would be saved in GraphML format. Default: nothing saved.
+        Where to save the extracted weight matrix. If networkx output is returned, it would be saved in GraphML format.
+        Default: nothing saved unless instructed.
 
     Returns
     -------
@@ -197,12 +218,12 @@ def extract(features, groups,
     """
 
     # parameter check
[…]
