### 2017-09-19 7be10a6
ignoring runtime exceptions
