+++ hiwenet/analyze_distance_behaviour.py
@@ -1,0 +1,208 @@
+"""
+
+Script to analyze the relationships among different histogram distances,
+    at different dimensionalities with different number of bins.
+
+"""
+import os
+import pickle
+from os.path import join as pjoin, exists as pexists
+
+import numpy as np
+from joblib import Parallel, delayed
+from matplotlib import pyplot as plt
+from medpy.metric import histogram
+
+metric_list = ['manhattan', 'minowski', 'euclidean', 'noelle_2', 'noelle_4', 'noelle_5']
+metric_list_to_doublecheck = ['cosine_1']
+unknown_property = ['histogram_intersection']
+still_under_dev  = ['quadratic_forms']
+similarity_funcs = ['correlate', 'cosine', 'cosine_2', 'cosine_alt', 'fidelity_based']
+semi_metric_list = ['jensen_shannon', 'chi_square', 'chebyshev', 'chebyshev_neg',
+                    'histogram_intersection_1', 'relative_deviation', 'relative_bin_deviation',
+                    'noelle_1', 'noelle_3', 'correlate_1']
+excluded = ['kullback_leibler', ]
+all_methods = metric_list + similarity_funcs + semi_metric_list
+all_methods.sort()
+
+default_feature_dim = 1000
+default_num_bins = 20
+default_stdev = 0.1
+
+range_dim = np.logspace(1, 3, num=20, dtype=int) # range(10, 100000, 1000)
+range_num_bins = np.array(list(range(5, 200, 20)))
+num_trials = 25
+
+num_methods = len(all_methods)
+num_dimensions = len(range_dim)
+num_num_bins = len(range_num_bins)
+
+num_rows = 3  # 7
+num_cols = 7  # 3
+fig_size = [14, 18]
+num_ticks = 10
+
+range_separability = [-0.75, -0.5, -0.25, 0.0, 0.25, 0.5, 0.75]
+center_one = 0.0
+
+edge_range = (-1.0, 1.0)
+common_bin_edges = lambda nb: np.linspace(edge_range[0], edge_range[1], nb)
+
+num_sep = len(range_separability)
+
+rng = np.random.RandomState(654321)
+
+def get_distr(center=0.0, stdev=default_stdev, length=50):
+    "Returns a PDF of a given length. "
+    # distr = np.random.random(length)
+
+    # sticking to normal distibution to easily control separability
+    distr = rng.normal(center, stdev, size=[length, 1])
+
+    return distr
+
+
+def make_random_histogram(center=0.0, stdev=default_stdev, length=default_feature_dim, num_bins=default_num_bins):
+    "Returns a sequence of histogram density values that sum to 1.0"
+
+    hist, bin_edges = np.histogram(get_distr(center, stdev, length),
+                                   range=edge_range, bins=num_bins, density=True)
+    # to ensure they sum to 1.0
+    hist = hist / sum(hist)
+
+    if len(hist) < 2:
+        raise ValueError('Invalid histogram')
+
+    return hist, bin_edges
+
+
+def dist_betn_rand_hist(method, center_one, center_two, feat_dim, num_bins):
+
+    h1, _ = make_random_histogram(center_one, length=feat_dim, num_bins=num_bins)
+    h2, _ = make_random_histogram(center_two, length=feat_dim, num_bins=num_bins)
+
+    return method(h1, h2)
+
+
+# -------------------
+
+print('range of dimensions : {}'.format(range_dim))
+
+out_dir = '/data1/strother_lab/praamana/dist_compare_hiwenet_lowerdim_finegrain'
+if not pexists(out_dir):
+    os.mkdir(out_dir)
+
+# print('Computation ...')
+# for separability in range_separability:
[…]
