### 2017-09-22 49994c6
printing all exceptions [skip ci]
