+++ docs/cite.md
@@ -1,6 +1,0 @@
-
-## Citation
-
-If you found it useful for your research, please cite it as:
-
- * Pradeep Reddy Raamana. (Version 2). Histogram-weighted Networks for Feature Extraction, Connectivity and Advanced Analysis in Neuroscience. Zenodo. http://doi.org/10.5281/zenodo.839995
