+++ setup.cfg
@@ -1,0 +1,8 @@
+
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = hiwenet/_version.py
+versionfile_build = hiwenet/_version.py
+tag_prefix =
+parentdir_prefix = hiwenet-
