+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include versioneer.py
+include hiwenet/_version.py
