+++ setup.py
@@ -22,6 +22,8 @@ setup(name='hiwenet',
               'Operating System :: POSIX',
               'Operating System :: Unix',
               'Operating System :: MacOS',
+              'Programming Language :: Python :: 3.5',
+              'Programming Language :: Python :: 3.6',              
               'Programming Language :: Python :: 2.7',
           ],
       entry_points={
