### 2017-10-17 77f0f06
Changing node labels from numeric indices to original group ids provided
