### 2017-10-17 2d136f8
including citations in the main paper
