### 2017-10-17 0405ead
including citations in the main paper itself
