### 2017-10-17 raamana/hiwenet#8 @raamana
Hey guys, sorry for the delay. I had been busy with other projects/deadlines.

Chris, I like your suggestions and have now added the references listed in paper.bib to paper.md. Let me know if I can improve it further.
