### 2017-10-17 9a8319e
improved readme and paper
