+++ README.md
@@ -11,84 +11,17 @@
 
 Histogram-weighted Networks for Feature Extraction and Advanced Analysis in Neuroscience
 
-Network-level analysis of various features, esp. if it can be individualized for a single-subject, is proving to be quite a valuable tool in many applications. This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node. Individual nodes could be an ROI or a patch or a cube, or any other unit of relevance in your application. This is a great way to take advantage of the full distribution of values available within each node, relative to the simpler use of averages (or another summary statistic). 
+Network-level analysis of various features, esp. if it can be individualized for a single-subject is proving to be a valuable tool in many applications. Ability to extract the networks for a given subject individually on its own, would allow for feature extraction conducive to predictive modeling, unlike group-wise networks which can only be used for descriptive and explanatory purposes. This package extracts single-subject (individualized, or intrinsic) networks from node-wise data by computing the edge weights based on histogram distance between the distributions of values within each node. Individual nodes could be an ROI or a patch or a cube, or any other unit of relevance in your application. This is a great way to take advantage of the full distribution of values available within each node, relative to the simpler use of averages (or another summary statistic) to compare two nodes/ROIs within a given subject.
 
 Rough scheme of computation is shown below:
 ![illustration](docs/illustration.png)
 
-## Applicability and target audience
-
-Although this technique was originally developed for cortical thickness, this is a generic and powerful technique that could be applied to any features such as gray matter density, PET uptake values, functional activation data or EEG features. All that is needed is a set of nodes/parcellation that have one-to-one correspondence across samples/subjects in your dataset.
-
-The target audience is users of almost all neuroimaging modalities who 1) preprocessed dataset already, 2) have some base features (node- or patch-wise, that are native to the given modality) extracted using other packages (metioned above), and 3) who would like to analyze network-level or covariance-type or connectivity relations among the base features.
-
 ## Installation
 
-`pip install hiwenet`
+`pip install -U hiwenet`
 
+## Documentation
 
-As you continue to use or run into issues, try `pip install -U hiwenet` to get the latest bug-fixes, new features and optimized performance. 
-
-#### verison compatiblity
-
-I plan to support all version of python. However, only python 2.7 is currently tested for support automatically as part of continuous integration. Support for Python 3+ will be coming shortly. 
-
-There is nothing in this package that is exotic for it to break across multiple versions of Python, however it's not automatically tested. If you would like to contribute to broaden compatibility, that's be appreciated.
-
-## Usage
-
-This package computes single-subject networks, hence you may need loop over samples/subjects in your dataset to extract them for all the samples/subjects, and them proceed to your subsequent analysis (such as classification etc).
-
-A rough example of usage can be:
-
-```python
-import hiwenet
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.model_selection import cross_val_score
-import numpy as np
-import nibabel
-import os
-
-# ------------------------------------------------------------------------------------
-# see docs/example_thickness_hiwenet.py for a concrete example
-# ------------------------------------------------------------------------------------
-
-for ss, subject in enumerate(subject_list):
-  features = get_features(subject)
-  edge_weight_matrix = hiwenet.extract(features, groups,  weight_method = 'kullback_leibler')
[…]
