### 2017-10-17 66d2ad8
improving docs
