+++ docs/metrics.rst
@@ -1,3 +1,4 @@
+-------
 Metrics
 -------
 
