+++ docs/guidelines.rst
@@ -1,0 +1,9 @@
+----------
+Guidelines
+----------
+
+How to choose a histogram metric and various parameters associated.
+
+Under construction.
+
+Stay tuned.
