### 2017-10-17 raamana/hiwenet#8 @raamana
Oscar, I believe I have addressed your comments as noted in https://github.com/openjournals/joss-reviews/issues/380#issuecomment-327850313 via revisions to the paper.md and README and other commits.

I've also added more analysis and [tests](https://github.com/raamana/hiwenet/blob/master/hiwenet/analyze_distance_behaviour.py) to analyze the behaviour of these distance metrics with changing distance/similarity. These are no mathematical proofs, but I am more confident than earlier that the package is doing what it is supposed to do. I am going to work on this further, as part of my research (and will be updating the package naturally) as I learn more about these distances. This would also partly address #10 as asserting the validity of current implementation is as or more important than checking whether it matches with my earlier matlab implementation.

Since we last chatted 6 weeks ago, I've improved the package in many ways, including implementing versioneer and option to control the range of histogram edges.

Updated docs, which are generated automatically, are at: http://hiwenet.readthedocs.io/

Take a look and let me know if these changes improve hiwenet and bring it closer to acceptance. Thank you again both for your time and comments to improve hiwenet.
