### 2017-10-17 raamana/hiwenet#4 @raamana
Hi Oscar, I believe #8 would fix this and let's centralize this discussion there if that's okay with you. Please reopen this otherwise.
