### 2017-10-25 raamana/hiwenet#8 @raamana
just to clarify, the [analysis](https://github.com/raamana/hiwenet/blob/master/hiwenet/analyze_distance_behaviour.py) I mentioned above, is in addition to the tests I wrote for the histogram distances specifically (ensuring their mathematical properties as metrics, whenever they are claimed to be). They've been found helpful and [merged](https://github.com/loli/medpy/pull/40) into the medpy package, and are also independently [run within `hiwenet`](https://github.com/raamana/hiwenet/blob/master/hiwenet/test_medpy.py) package itself for double checking.
