### 2017-11-08 387b1e3
zapping leading double underscores in method names to improve readability
