+++ hiwenet/hiwenet.py
@@ -47,7 +47,7 @@ default_trim_behaviour = True
 default_return_networkx_graph = False
 default_out_weights_path = None
 
-def __compute_bin_edges(features, num_bins, edge_range, trim_outliers, trim_percentile):
+def compute_bin_edges(features, num_bins, edge_range, trim_outliers, trim_percentile):
     "Compute the edges for the histogram bins to keep it the same for all nodes."
 
     if edge_range is None:
@@ -219,12 +219,12 @@ def extract(features, groups,
     """
 
     # parameter check
-    features, groups, num_bins, edge_range, weight_method, group_ids, num_groups, num_links = __parameter_check(
+    features, groups, num_bins, edge_range, weight_method, group_ids, num_groups, num_links = check_params(
         features, groups, num_bins, edge_range, weight_method, trim_outliers, trim_percentile)
 
     # using the same bin edges for all nodes/groups to ensure correspondence
     # NOTE: common bin edges is important for the disances to be any meaningful
-    edges = __compute_bin_edges(features, num_bins, edge_range, trim_outliers, trim_percentile)
+    edges = compute_bin_edges(features, num_bins, edge_range, trim_outliers, trim_percentile)
 
     if return_networkx_graph:
         nx_graph = nx.Graph()
@@ -237,14 +237,14 @@ def extract(features, groups,
         if np.mod(g1 + 1, 5) == 0.0:
             sys.stdout.write('.')
         index1 = groups == group_ids[g1]
-        hist_one = __compute_histogram(features[index1], edges)
+        hist_one = compute_histogram(features[index1], edges)
 
         for g2 in range(g1 + 1, num_groups, 1):
             index2 = groups == group_ids[g2]
-            hist_two = __compute_histogram(features[index2], edges)
+            hist_two = compute_histogram(features[index2], edges)
 
             try:
-                edge_value = _compute_edge_weight(hist_one, hist_two, weight_method)
+                edge_value = compute_edge_weight(hist_one, hist_two, weight_method)
                 if return_networkx_graph:
                     nx_graph.add_edge(group_ids[g1], group_ids[g2], weight=float(edge_value))
                 else:
@@ -273,16 +273,16 @@ def extract(features, groups,
         return edge_weights
 
 
-def __compute_histogram(values, edges):
+def compute_histogram(values, edges):
     """Computes histogram (density) for a given vector of values."""
 
     hist, bin_edges = np.histogram(values, bins=edges, density=True)
-    hist = __preprocess_histogram(hist, values, edges)
+    hist = preprocess_histogram(hist, values, edges)
 
     return hist
 
 
-def __preprocess_histogram(hist, values, edges):
+def preprocess_histogram(hist, values, edges):
     """Handles edge-cases and extremely-skewed histograms"""
 
     # working with extremely skewed histograms
@@ -297,7 +297,7 @@ def __preprocess_histogram(hist, values, edges):
     return hist
 
 
-def _compute_edge_weight(hist_one, hist_two, weight_method_str):
+def compute_edge_weight(hist_one, hist_two, weight_method_str):
     """
     Computes the edge weight between the two histograms.
 
@@ -325,7 +325,7 @@ def _compute_edge_weight(hist_one, hist_two, weight_method_str):
     return edge_value
 
 
-def _identify_groups(groups):
+def identify_groups(groups):
     """
     To compute number of unique elements in a given membership specification.
 
@@ -353,7 +353,7 @@ def _identify_groups(groups):
     return group_ids, num_groups
 
 
-def _range_check_parameters(num_bins, num_groups, num_values, trim_outliers, trim_percentile):
+def check_param_ranges(num_bins, num_groups, num_values, trim_outliers, trim_percentile):
[…]
