### 2017-11-08 ff84e1f
new feature: allowing input of arbitrary callable (only API, not CLI)
