### 2017-11-08 03c5ebf
new feature: allowing user function on orig data (without building histogrms)
