### 2017-11-12 9868556
changing default value to NaN to prevent unintended usage
