### 2017-11-12 7cf61bb
new feature: allowing non-symmetric adjacency matrices
