+++ hiwenet/hiwenet.py
@@ -37,6 +37,8 @@ semi_metric_list = [
     'noelle_1', 'noelle_3',
     'correlate_1']
 
+metrics_on_original_features = ['diff_medians', ]
+
 minimum_num_bins = 5
 
 default_weight_method = 'manhattan'
@@ -78,6 +80,7 @@ def extract(features, groups,
             trim_outliers=default_trim_behaviour,
             trim_percentile=default_trim_percentile,
             use_original_distribution=False,
+            asymmetric=False,
             return_networkx_graph=default_return_networkx_graph,
             out_weights_path=default_out_weights_path):
     """
@@ -104,12 +107,6 @@ def extract(features, groups,
     weight_method : string or callable, optional
         Type of distance (or metric) to compute between the pair of histograms.
         It can either be a string identifying one of the weights implemented below, or a valid callable.
-
-        If a callable, it must two accept two arrays as input and return one scalar as output.
-            Example: ``diff_in_skew = lambda x, y: abs(scipy.stats.skew(x)-scipy.stats.skew(y))``
-            NOTE: this method will be applied to histograms (not the original distribution of features from group/ROI).
-            In order to apply this callable directly on the original distribution (without trimming and histogram binning),
-            use ``use_original_distribution=True``.
 
         If a string, it must be one of the following methods:
 
@@ -173,6 +170,22 @@ def extract(features, groups,
 
         *Default* choice: 'minowski'.
 
+        The method can also be one of the following identifying metrics that operate on the original data directly -
+         e.g. difference in the medians coming from the distributions of the pair of ROIs.
+
+         - 'diff_medians'
+         - 'diff_means'
+
+         Please note this can lead to adjacency matrices that may not be symmetric
+            e.g. difference metric on two scalars is not symmetric).
+            In this case, be sure to use the flag: allow_non_symmetric=True
+
+        If weight_method is a callable, it must two accept two arrays as input and return one scalar as output.
+            Example: ``diff_in_skew = lambda x, y: abs(scipy.stats.skew(x)-scipy.stats.skew(y))``
+            NOTE: this method will be applied to histograms (not the original distribution of features from group/ROI).
+            In order to apply this callable directly on the original distribution (without trimming and histogram binning),
+            use ``use_original_distribution=True``.
+
     num_bins : scalar, optional
         Number of bins to use when computing histogram within each patch/group.
 
@@ -217,6 +230,11 @@ def extract(features, groups,
         This option is valid only when weight_method is a valid callable,
             which must take two inputs (possibly of different lengths) and return a single scalar.
 
+    asymmetric : bool
+        Flag to identify resulting adjacency matrix is expected to be non-symmetric.
+        Note: this results in twice the computation time!
+        Default: False , for histogram metrics implemented here are symmetric.
+
     return_networkx_graph : bool, optional
         Specifies the need for a networkx graph populated with weights computed. Default: False.
 
@@ -245,35 +263,47 @@ def extract(features, groups,
     features, groups, num_bins, edge_range, group_ids, num_groups, num_links = check_params(
         features, groups, num_bins, edge_range, trim_outliers, trim_percentile)
 
-    weight_func, use_orig_distr = check_weight_method(weight_method, use_original_distribution)
+    weight_func, use_orig_distr, non_symmetric = check_weight_method(weight_method, use_original_distribution, asymmetric)
 
     # using the same bin edges for all nodes/groups to ensure correspondence
     # NOTE: common bin edges is important for the disances to be any meaningful
[…]
