### 2017-11-12 54486e3
new metrics: difference in medians
