### 2017-11-12 raamana/hiwenet#8 @raamana
Additional non-histogram metrics are implemented via #12.
