### 2017-11-12 raamana/hiwenet#8 @raamana
Also, ability to build directed networks (asymmetric adjacency matrices) is added!

Via 7cf61bb
