### 2017-11-12 2908669
ignoring NaNs/Non-finite values via masked arrays; new metric: diff in means
