### 2017-11-12 b3b0d92
recognizing when the metric is symmetric, to save time for user!
