### 2017-11-18 d484bcf
adding .bettercodehub.yml [skip ci]
