+++ .bettercodehub.yml
@@ -2,6 +2,8 @@ exclude:
 - /docs/.*
 - /examples/simple_visualization_examples.py
 - /hiwenet/_version.py
+- /setup.py
+- /versioneer.py
 component_depth: 1
 languages:
 - python
