### 2017-11-20 raamana/hiwenet#8 @raamana
Hey Oscar, Chris, I know you guys are extremely busy, if you could take a quick look and let me know if you are happy with the revisions and latest state, or identify what else you would like, that would be great. I am hoping to tick development of this repo off my list soon.

Thanks for your time and expertise in reviewing and improving this repo so far. It's much improved since its submission.
