### 2017-11-20 raamana/hiwenet#8 @oesteban
Hi, I assumed that once the review is done, it was just a matter of fixing all the issues for the submitter. I could be wrong.

Anyways, I totally trust your criteria to close the review issues, but if you want me to double check or help I'm happy to come back to this.
