### 2017-11-20 raamana/hiwenet#8 @raamana
Thanks Oscar. I believe only the reviewer or editor is supposed to tick them off, once they believe the revisions are done and adequate. Here is the link : https://github.com/openjournals/joss-reviews/issues/380
