### 2017-11-21 raamana/hiwenet#10 @oesteban
I believe it would make sense to extend the added test to all metrics, but I leave this decision to @raamana.

I also consider that test 2 is not that hard to implement (it just takes pre-calculating the features file on a certain dataset with the MATLAB version and check that it is close enough to the features calculated from the same dataset with python, and the dataset does not need to come from a real brain, they could just be randomly generated). This test would add real value to the testing effort. As I said, it is not something I consider mandatory, so I'll close this issue.
