### 2017-11-21 #380 @oesteban
I think this is finished.
