### 2017-11-21 raamana/hiwenet#10 @raamana
Thanks Oscar, its definitely to run the CLI vs API matching test to all metrics, I haven't done, as I believe if it does for one, it should also behave the same for others, and I didn't want to elongate the tests for no obvious reasons. I will extend them.

Regarding matlab vs python, as I mentioned before, I expect them to differ as they are implemented by different people , possibly with different equations, although being in the same family. Moreover, i don't any consider any to be a reference or ground truth - each are simply a different instantiation (although with different implementations). I agree with you doing this is way better than otherwise, I hope to do it when I am able to, or when I am able to find other resources (like a coop project).
