### 2017-11-21 fb224f2
extending CLI vs API test to all metrics and semi metrics
