### 2017-11-21 raamana/hiwenet#10 @raamana
CLI vs API matching test has been extended - via fb224f23f03b2e537fb94027f0b2a84ded836d44 -  to all metrics and semi metrics now, thanks for pointing it out.
