### 2017-11-21 6ecc394
extended test compatibility for python 2.7
