### 2017-11-21 acc843f
Clean up minor changes
