+++ CHANGELOG.md
@@ -1,0 +1,9 @@
+# Changelog
+
+## 0.4 (2017-11-21)
+
+- New feature: ability to apply arbitrary [distance/similarity] functions
+- New feature: ability to apply functions on original data.
+- New metrics: difference in medians [and means]
+- Better handling of NaNs/Non-finite values via masked arrays
+- Recognizing when the metric symmetric/not
