### 2017-11-21 #380 @raamana
Here is the new final DOI: 10.5281/zenodo.1064012

Thank a lot Oscar and Chris, your review and suggestions helped improve the software greatly.
