### 2017-11-21 18dc117
Updating the citation info with latest DOI [skip ci]
