+++ docs/citation.rst
@@ -1,0 +1,10 @@
+Citation
+--------
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1064012.svg)](https://doi.org/10.5281/zenodo.1064012)
+
+If you found this toolbox useful for your research, please cite one or more of these papers in that order:
+
+ - Raamana, P. R. and Strother, S.C., 2017, Histogram-weighted Networks for Feature Extraction, Connectivity and Advanced Analysis in Neuroscience. Zenodo. https://doi.org/10.5281/zenodo.1064012
+ - Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381.
+- Raamana, P. R., Weiner, M. W., Wang, L., Beg, M. F., & Alzheimer's Disease Neuroimaging Initiative. (2015). Thickness network features for prognostic applications in dementia. Neurobiology of aging, 36, S91-S102.
