### 2017-11-26 d95b11b
new module to implement non-pairwise metric: relative_to_all
