+++ hiwenet/__init__.py
@@ -4,15 +4,16 @@
 
 """
 
-__all__ = ['extract', 'pairwise_dist', 'run_cli', 'more_metrics']
+__all__ = ['extract', 'pairwise_dist', 'run_cli', 'more_metrics', 'non_pairwise']
 
 from sys import version_info
 
 if version_info.major==2 and version_info.minor==7:
     import more_metrics
+    import non_pairwise
     from pairwise_dist import extract, run_cli
 elif version_info.major > 2:
-    from hiwenet import more_metrics
+    from hiwenet import more_metrics, non_pairwise
     from hiwenet.pairwise_dist import extract, run_cli
 else:
     raise NotImplementedError('hiwenet supports only 2.7 or 3+. Upgrade to Python 3+ is recommended.')
