### 2017-11-26 d6b10fd
accommodating the special metric
