### 2017-11-26 6a724f2
code cleanup
