### 2017-11-26 1768bdf
refactoring the core function to make the publicly exposed function more readable
