### 2017-11-26 da369c5
collecting common methods as utilities
