### 2017-11-26 90759ce
randomizing the size of groups and dimensionalities - more stringent tests
