### 2017-11-26 #380 @raamana
This DOI corresponds to the [latest GitHub release](https://github.com/raamana/hiwenet/releases/tag/0.4) tagged 0.4.

Also wouldn't JOSS be providing another DOI upon publication?
