### 2017-11-26 #380 @cMadan
@raamana, perfect! I've set the archive DOI and updated the version number. The archive DOI is for the software release, whereas the JOSS DOI will be for the paper. (There are also different kinds of DOIs--DataCite vs. CrossRef--which have differences in how their associated metadata and how they are processed by services like Google Scholar and PubMed.)

@arfon, we're all set to accept here!
