### 2017-11-27 e94b9a1
Update paper.md
