### 2017-11-27 #380 @raamana
The PDF produced doesn't look right - image not aligned. I've updated it to try more space around it, and improved the text a bit more. If possible, it may be better to try generate the PDF again.
