### 2017-11-27 #380 @arfon
@raamana - the figure has now flowed onto the next page (which might be OK?). If this works for you, perhaps you should just change the previous sentence to 'Rough scheme of computation shown in Figure 1'
