### 2017-11-27 bbfc84d
more generic reference to figure
