### 2017-11-27 #380 @arfon
@raamana - your submission is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00380 ⚡️ 🚀 💥
